Update labOneClickDeployment.ps1
I was receiving an error "New-AzStorageAccount : The storage account named ......stor is already taken".  I noticed that based on the documentation of what it was supposed to be naming it was cutting off the end of the name and only appending "stor".  This lead to finding the character length for storageAccountName being defined as (0,22) where Azure defines it as being 24 in length.  Modified code to reflect.  Not sure if it should have been changed on all lines for all variables but I did if someone can verify...lines 136, 138, 143, 145.  This allowed the script to run and the necessary account and containers to be created.